Move CountProxy
