+++ src/Support/DataStructure/Set.php
@@ -4,6 +4,7 @@ namespace ArtARTs36\MergeRequestLinter\Support\DataStructure;
 
 use ArtARTs36\MergeRequestLinter\Contracts\Collection;
 use ArtARTs36\MergeRequestLinter\Support\ArrayKeyIterator;
+use ArtARTs36\MergeRequestLinter\Support\DataStructure\Traits\CountProxy;
 
 /**
  * @template V
